# docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      error:
        example: Error message here
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      message:
        example: Login successful
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/models.UserSummary'
    type: object
...
  models.SignupRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      phone:
        example: "+1234567890"
        type: string
      user_type:
        enum:
        - USER
        - ADMIN
        example: USER
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - phone
    - user_type
    type: object
  models.SignupResponse:
    properties:
      message:
        example: User created successfully
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.User:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        maxLength: 100
        minLength: 2
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      last_name:
        example: Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      phone:
        example: "+1234567890"
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_type:
        enum:
        - USER
        - ADMIN
        example: USER
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - phone
    - user_type
    type: object
  models.UserSummary:
    properties:
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      last_name:
        example: Doe
        type: string
      user_type:
        enum:
        - USER
        - ADMIN
        example: USER
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@restaurant.com
    name: API Support
  description: This is a restaurant management system API with user authentication,
    food ordering, and invoice management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns JWT access token
      parameters:
      - description: Email and Password
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with authentication token and user details
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User Login
      tags:
      - Authentication
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user account with email, password, and profile information
      parameters:
      - description: User Registration Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully with authentication token
          schema:
            $ref: '#/definitions/models.SignupResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Email already exists in the system
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User Signup
      tags:
      - Authentication
  /auth/user:
    get:
      consumes:
      - application/json
      description: Get authenticated user profile details (requires valid JWT token)
      produces:
      - application/json
      responses:
        "200":
          description: User profile details (password field will be empty)
          schema:
            $ref: '#/definitions/models.UserSummary'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found in database
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Current User
      tags:
      - Authentication
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"